requirements.txt
PyYAML
docxtpl
Pillow
pyinstaller
